Get segment ids of every faces from the polyhedron
After partitioning in the original source file it is unclear how to get partition ids of faces from polyhedron to distinguish the faces according to segments they belong. In the added code the parameter 'segmId' gets access to that information and it can be used, for example, in coloring polyhedron according to its segments or parts.